algunos cambios sobre la barra de vida
quiero guardar estas cosas porque estoy por probar a ver si funciona algo (cabe destacar que como está ahora no funciona del todo así que no  lo tomes en cuenta) :3

+++ Diagony/Assets/Scripts/CombatScene/Enemys/Ira.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Ira : MonoBehaviour
 {
-    public float Health = 100;
-    public float MaxHealth = 100;
+    public int Health = 100;
+    public int MaxHealth = 100;
     public int Atq = 10;
     void Start()
     {
